+++ themes/material/src/classes/disabled_cls.rs
@@ -1,12 +1,11 @@
 use ribir_core::prelude::*;
 
 pub(super) fn init(classes: &mut Classes) {
-  classes.insert(DISABLED, |w| {
-    color_filter! {
-      filter: GRAYSCALE_FILTER,
+  classes.insert(
+    DISABLED,
+    style_class! {
+      filter: Filter::grayscale(1.),
       opacity: 0.38,
-      @ { w }
-    }
-    .into_widget()
-  });
+    },
+  );
 }